Add mapping between ISO language code and friendly name [#33786827]
Friendly language names are now displayed throughout the UI while the
ISO codes are used when communicating with the backend. Twitter has a
list of supported languages at
http://api.twitter.com/1/help/languages.json but it appears this is out
of sync with actual languages in the searh results. It's trivial to
support more so I will fix them when I Come across them.